use maven to build project IDE independently
