--- .classpath
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.USER_LIBRARY/ANT"/>
-	<classpathentry kind="lib" path="lib/ant-contrib-0.6.jar"/>
-	<classpathentry kind="lib" path="lib/batik-css-1.7.jar"/>
-	<classpathentry kind="lib" path="lib/batik-util-1.7.jar"/>
-	<classpathentry kind="lib" path="lib/sac-1.3.jar"/>
-	<classpathentry kind="lib" path="lib/epubcheck-3.0.1.jar"/>
-	<classpathentry kind="lib" path="lib/commons-compress-1.4.1.jar"/>
-	<classpathentry kind="lib" path="lib/jing-20120724.0.0.jar"/>
-	<classpathentry kind="lib" path="lib/guava-13.0.1.jar"/>
-	<classpathentry kind="lib" path="lib/Saxon-HE-9.4.0.6.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
