--- in/.gitignore
@@ -1,5 +1,0 @@
-/artistesfous/
-/lamado/
-/madtelier/
-/misc/
-/sema/
